fix(bridge-flows): review findings (4)
Indexer
-------
• MessageAttestedTo was not idempotent: an event replay (Envio restart,
  chain reorg) would increment attestationCount each time even though
  the BridgeAttestation row is unique-keyed. Now the handler probes the
  attestation row first and skips the BridgeTransfer counter bump when
  the attestation already exists.

• Removed the QUEUED_OUTBOUND status branch and its unused detail
  fields (outboundQueuedSequence, rateLimitedCurrentCapacity). The
  branch was unreachable: source-side queue events fire with only a
  sequence, not a digest, so they can't be correlated to BridgeTransfer
  without a sequence→digest lookup. Dropped the three no-op handlers,
  the YAML event bindings, and the ABI entries. QUEUED_INBOUND (dest-
  side, digest-keyed) stays — it's populated and reachable.

• Added @index on BridgeTransfer.firstSeenAt. Both BRIDGE_TRANSFERS_WINDOW
  and BRIDGE_TRANSFER_COUNT filter+sort on it (since the dest-first-race
  fix switched away from sentTimestamp), so the missing index would
  force full-table scans under load.

Refactor
--------
• Extracted defaultWormholeDetail and moved bytes32ToAddress into a new
  src/wormhole/detail.ts. Collapses the ~10-line default object that
  was duplicated across nttManager.ts and wormholeTransceiver.ts, and
  keeps the generic src/bridge.ts free of Wormhole-specific helpers.

Schema + UI kept consistent: types.ts, bridge-status.ts, bridge-queries.ts
dropped the same three fields and QUEUED_OUTBOUND status.
