+++ ui-dashboard/src/lib/bridge-status.ts
@@ -25,7 +25,6 @@ const STATUS_CLASSES: Record<BridgeStatusOverlay, string> = {
   SENT: "bg-slate-800 text-slate-300",
   ATTESTED: "bg-indigo-900/40 text-indigo-300",
   DELIVERED: "bg-emerald-900/40 text-emerald-300",
-  QUEUED_OUTBOUND: "bg-amber-900/40 text-amber-300",
   QUEUED_INBOUND: "bg-amber-900/40 text-amber-300",
   CANCELLED: "bg-slate-800 text-slate-500",
   FAILED: "bg-red-900/60 text-red-200",
@@ -37,8 +36,7 @@ const STATUS_LABELS: Record<BridgeStatusOverlay, string> = {
   SENT: "Sent",
   ATTESTED: "Attested",
   DELIVERED: "Delivered",
-  QUEUED_OUTBOUND: "Queued (out)",
-  QUEUED_INBOUND: "Queued (in)",
+  QUEUED_INBOUND: "Queued",
   CANCELLED: "Cancelled",
   FAILED: "Failed",
   STUCK: "Stuck",
